Updated scviol link
Old link was dead, I found the the file in the new place and updated it. I made it a permanent link, which means that it will point to that version of the file forever but wont break, let me know if you want me to update it to a plain old link.