Proposed change for presentation date
I just realized that a paper deadline that I may be aiming for is on March 17, so as much as I would have loved to lead the discussion on alias analysis, I think it's best for me to reschedule. Sorry @andrewb1999 for the false alarm! And @5hubh4m I hope you're cool with me joining you :slightly_smiling_face: